+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added airlift json source and airlift json transform to vector.toml ([#553]).
+
+[#553]: https://github.com/stackabletech/operator-rs/pull/553
+
 ## [0.34.0] - 2023-02-06
 
 ### Added